Move compute_diameter functions in one place.
